requirements.txt
Pillow==6.1.0
tqdm==4.32.1
requests==2.22.0